requirements.txt
streamlit
pinecone-client
langchain-pinecone
langchain-google-genai
datasets
requests
python-dotenv
